+++ R/modelsummary.R
@@ -745,7 +745,7 @@ modelsummary <- function(
   # stars
   stars_note <- settings_get("stars_note")
   if (isTRUE(stars_note) && !isFALSE(stars) && !any(grepl("\\{stars\\}", c(estimate, statistic)))) {
-    stars_note <- make_stars_note(stars, output_format = output_format)
+    stars_note <- make_stars_note(stars, output_format = output_format, output_factory = output_factory)
     if (is.null(notes)) {
       notes <- stars_note
     } else {