+++ NEWS.md
@@ -1,5 +1,11 @@
 # News 
 
+## Development
+
+Bugs:
+
+* Stars note in `kableExtra` escaped `\num{}` in LaTeX.
+
 ## 2.2.0
 
 * Bump minimum version requirement for `tinytable`, `parameters`, and `insight` dependencies.